Fix issue where intellect treats bad filepath as actual rules and fails to raise any exceptions.
Intellect will raise a generic exception if it cannot learn a string when treated as a path to a file, or as actual rules. It may be appropriate to write custom exception classes to handle these types of errors. Previously, strings were examined as actual rules first, then examined as paths to files. The opposite is now true.